Moved tests to single class
There is still some overlap between tested methods but fixing this is low priority at this time. At least, the same coverage is in the same test class
